chore: use a different module
